# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ecom mini app",
        "contact": {}
    },
    "paths": {
        "/item": {
            "get": {
                "description": "Возвращает массив товара",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Получить весь товар",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ResponseCustom"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Items"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Items not found!",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Возвращает ID созданного товара",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Создать товар",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateItemInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ResponseCustom"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.ResponseID"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Can`t create item",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/item/{id}": {
            "get": {
                "description": "Возвращает товар с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Получить один товар",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ResponseCustom"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Items"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Item not found!",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет товар по id и возвращает флаг true",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Удалить товар",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ResponseCustom"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "error\": \"error\": \"Can`t delete item",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет товар по ID и возвращает ID этого товара",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "Обновить товар",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateItemInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ResponseCustom"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.ResponseID"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Can`t update iten",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
        "controllers.ResponseCustom": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "controllers.ResponseID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtos.CreateItemInput": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "price": {
                    "type": "number",
                    "minimum": 1
                }
            }
        },
...
        "dtos.UpdateItemInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "price": {
                    "type": "number",
                    "minimum": 1
                }
            }
        },
...
        "models.Items": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}